based crud setup for custom actions
