+++ apps/custom_actions/tables.py
@@ -1,0 +1,34 @@
+from django.conf import settings
+from django_tables2 import tables
+
+from apps.custom_actions.models import CustomAction
+from apps.generics import actions
+
+
+class CustomActionTable(tables.Table):
+    name = tables.columns.Column(
+        linkify=True,
+        attrs={
+            "a": {"class": "link"},
+        },
+        orderable=True,
+    )
+    actions = actions.ActionsColumn(
+        actions=[
+            actions.edit_action(
+                "custom_actions:edit",
+                required_permissions=["custom_actions.change_customaction"],
+            ),
+            actions.delete_action(
+                "custom_actions:delete",
+                required_permissions=["custom_actions.delete_customaction"],
+            ),
+        ]
+    )
+
+    class Meta:
+        model = CustomAction
+        fields = ("name", "description")
+        row_attrs = settings.DJANGO_TABLES2_ROW_ATTRS
+        orderable = False
+        empty_text = "No actions found."
