+++ apps/custom_actions/views.py
@@ -1,68 +1,15 @@
-from django import forms
-from django.conf import settings
 from django.contrib import messages
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.urls import reverse
 from django.views import View
 from django.views.generic import CreateView, TemplateView, UpdateView
-from django_tables2 import SingleTableView, tables
+from django_tables2 import SingleTableView
 
-from apps.custom_actions.fields import JSONORYAMLField
+from apps.custom_actions.forms import CustomActionForm
 from apps.custom_actions.models import CustomAction
-from apps.generics import actions
+from apps.custom_actions.tables import CustomActionTable
 from apps.teams.mixins import LoginAndTeamRequiredMixin
-
-
-class CustomActionForm(forms.ModelForm):
-    description = forms.CharField(widget=forms.Textarea(attrs={"rows": "3"}), required=False)
-    prompt = forms.CharField(
-        widget=forms.Textarea(attrs={"rows": "3"}),
-        required=False,
-        label="Additional Prompt",
-        help_text="Use this field to provide additional instructions to the LLM",
-    )
-    api_schema = JSONORYAMLField(
-        widget=forms.Textarea(attrs={"rows": "10"}),
-        required=True,
-        label="API Schema",
-        help_text="Paste in the OpenAPI schema for the API you want to interact with. "
-        "This will be used to generate the API calls for the LLM. Accepts YAML or JSON.",
-        initial={},
-    )
-
-    class Meta:
-        model = CustomAction
-        fields = ("name", "description", "prompt", "api_schema")
-
-
-class CustomActionTable(tables.Table):
-    name = tables.columns.Column(
-        linkify=True,
-        attrs={
-            "a": {"class": "link"},
-        },
-        orderable=True,
-    )
-    actions = actions.ActionsColumn(
-        actions=[
-            actions.edit_action(
-                "custom_actions:edit",
-                required_permissions=["custom_actions.change_customaction"],
-            ),
-            actions.delete_action(
-                "custom_actions:delete",
-                required_permissions=["custom_actions.delete_customaction"],
-            ),
-        ]
-    )
-
-    class Meta:
-        model = CustomAction
-        fields = ("name", "description")
-        row_attrs = settings.DJANGO_TABLES2_ROW_ATTRS
-        orderable = False
-        empty_text = "No actions found."
 
 
 class CustomActionHome(LoginAndTeamRequiredMixin, TemplateView):
